+++ app/src/features/surveys/view/SurveyHeader.test.tsx
@@ -44,9 +44,6 @@ const mockSurveyContext: ISurveyContext = {
   artifactDataLoader: {
     data: null
   } as DataLoader<any, any, any>,
-  sampleSiteDataLoader: {
-    data: null
-  } as DataLoader<any, any, any>,
   critterDataLoader: {
     data: null
   } as DataLoader<any, any, any>,
