+++ app/src/features/surveys/view/SurveyPage.tsx
@@ -7,7 +7,7 @@ import { SurveyContext } from 'contexts/surveyContext';
 import { TaxonomyContextProvider } from 'contexts/taxonomyContext';
 import SurveyDetails from 'features/surveys/view/SurveyDetails';
 import React, { useContext, useEffect } from 'react';
-import { SurveySamplingContainer } from './components/sampling-data/SurveySamplingContainer';
+import { SurveySamplingTableContainer } from './components/sampling-data/SurveySamplingTableContainer';
 import SurveyStudyArea from './components/SurveyStudyArea';
 import { SurveySpatialContainer } from './survey-spatial/SurveySpatialContainer';
 import SurveyAttachments from './SurveyAttachments';
@@ -36,7 +36,7 @@ const SurveyPage: React.FC = () => {
       <Container maxWidth="xl" sx={{ py: 3 }}>
         <Stack gap={3}>
           <Paper>
-            <SurveySamplingContainer />
+            <SurveySamplingTableContainer />
           </Paper>
 
           <Paper>
