+++ app/src/features/surveys/view/components/SurveyStudyArea.test.tsx
@@ -42,7 +42,6 @@ describe.skip('SurveyStudyArea', () => {
   it('renders correctly with no data', async () => {
     const mockSurveyDataLoader = { data: getSurveyForViewResponse } as DataLoader<any, any, any>;
     const mockArtifactDataLoader = { data: null } as DataLoader<any, any, any>;
-    const mockSampleSiteDataLoader = { data: null } as DataLoader<any, any, any>;
     const mockCritterDataLoader = { data: [] } as DataLoader<any, any, any>;
     const mockTechniqueDataLoader = { data: [] } as DataLoader<any, any, any>;
 
@@ -54,7 +53,7 @@ describe.skip('SurveyStudyArea', () => {
           surveyDataLoader: mockSurveyDataLoader,
           artifactDataLoader: mockArtifactDataLoader,
           techniqueDataLoader: mockTechniqueDataLoader,
-          sampleSiteDataLoader: mockSampleSiteDataLoader,
+          
           critterDataLoader: mockCritterDataLoader
         }}>
         <SurveyStudyArea />
@@ -89,7 +88,7 @@ describe.skip('SurveyStudyArea', () => {
             surveyId: 1,
             surveyDataLoader: mockSurveyDataLoader,
             artifactDataLoader: mockArtifactDataLoader,
-            sampleSiteDataLoader: mockSampleSiteDataLoader,
+            
             techniqueDataLoader: mockTechniqueDataLoader,
             critterDataLoader: mockCritterDataLoader
           }}>
@@ -117,7 +116,7 @@ describe.skip('SurveyStudyArea', () => {
             surveyId: 1,
             surveyDataLoader: mockSurveyDataLoader,
             artifactDataLoader: mockArtifactDataLoader,
-            sampleSiteDataLoader: mockSampleSiteDataLoader,
+            
             critterDataLoader: mockCritterDataLoader,
             techniqueDataLoader: mockTechniqueDataLoader
           }}>
@@ -138,7 +137,6 @@ describe.skip('SurveyStudyArea', () => {
       refresh: jest.fn() as unknown as any
     } as DataLoader<any, any, any>;
     const mockArtifactDataLoader = { data: null } as DataLoader<any, any, any>;
-    const mockSampleSiteDataLoader = { data: null } as DataLoader<any, any, any>;
     const mockCritterDataLoader = { data: [] } as DataLoader<any, any, any>;
     const mockTechniqueDataLoader = { data: [] } as DataLoader<any, any, any>;
 
@@ -159,7 +157,7 @@ describe.skip('SurveyStudyArea', () => {
             surveyId: 1,
             surveyDataLoader: mockSurveyDataLoader,
             artifactDataLoader: mockArtifactDataLoader,
-            sampleSiteDataLoader: mockSampleSiteDataLoader,
+            
             critterDataLoader: mockCritterDataLoader,
             techniqueDataLoader: mockTechniqueDataLoader
           }}>
@@ -230,7 +228,6 @@ describe.skip('SurveyStudyArea', () => {
   it('shows error dialog with API error message when updating survey data fails', async () => {
     const mockSurveyDataLoader = { data: getSurveyForViewResponse } as DataLoader<any, any, any>;
     const mockArtifactDataLoader = { data: null } as DataLoader<any, any, any>;
-    const mockSampleSiteDataLoader = { data: null } as DataLoader<any, any, any>;
     const mockCritterDataLoader = { data: [] } as DataLoader<any, any, any>;
     const mockTechniqueDataLoader = { data: [] } as DataLoader<any, any, any>;
 
@@ -269,7 +266,7 @@ describe.skip('SurveyStudyArea', () => {
             surveyId: 1,
             surveyDataLoader: mockSurveyDataLoader,
             artifactDataLoader: mockArtifactDataLoader,
-            sampleSiteDataLoader: mockSampleSiteDataLoader,
+            
             critterDataLoader: mockCritterDataLoader,
             techniqueDataLoader: mockTechniqueDataLoader
           }}>
